Refactor identification method
